Allow colons in filenames
Due to the ambiguous nature of the output format is impossible to make it work
in all the cases, since both the path and the matched line can contain arbitrary
characters.

Looking for positive numbers surrounded by colon, instead of just a colon,
allows to make it work in a much wider variety of cases; it still can't handle
filenames like `foo:1:bar.baz`.